requirements.txt
requests
aiohttp
asyncio
google-generativeai
python-dotenv
mermaid-py
playwright
boto3
fastapi
uvicorn
certifi